% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barplot.leem}
\alias{barplot.leem}
\title{Barplot graph}
\usage{
\method{barplot}{leem}(
  x,
  bg = TRUE,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  panel.first = grid(col = "white"),
  bgcol = "gray",
  bgborder = NA,
  barcol = "yellow",
  barborder = "gray",
  posx1 = 0,
  posx2 = 0,
  xangf = 0,
  labels = NULL,
  ...
)
}
\description{
Barplot graph
}
