% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert2.R
\name{insert}
\alias{insert}
\title{Insert measures of position in plot}
\usage{
insert(x, ...)
}
\arguments{
\item{x}{R object (list) of class leem. Use \code{new_leem()} function.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function that allows inserting measures of position in plots
}
\examples{
# Example 1
library(leem)
set.seed(10)
rnorm(36, 100, 50) |>
 new_leem(variable = "continuous") |>
 tabfreq() |>
 hist() |>
 insert(
  lcol = "black",
  tcol = "purple",
  acol = "brown",
  parrow = 0.6,
  larrow = 0.6,
  ptext = 0.4,
  side = "left",
  lwd = 2,
  lwdarrow = 4
 )
}
