% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{leem}
\alias{leem}
\title{Graphical User Interface for leem package}
\usage{
leem(gui = TRUE)
}
\arguments{
\item{gui}{Logical argument, \code{TRUE} or \code{FALSE}. The default is \code{TRUE}}
}
\value{
\code{leem} presents GUI with various problems for the teaching of statistics and mathematics. The idea is to use this package to learn these subjects without necessarily programming in R
}
\description{
\code{leem} A Graphical User Interface (GUI) for
    the leem package
}
\examples{
# Loading package
library(leem)
if (interactive()) {
  leem(gui = FALSE)
}
}
