% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newleem.R
\name{new_leem}
\alias{new_leem}
\title{Constructor of object of leem class}
\usage{
new_leem(x = vector(), variable = "discrete")
}
\arguments{
\item{x}{R object (vector as data structure).}

\item{variable}{Type of data. If \code{discrete} (default), the data are categorical (numeric or not). If continuous, the data are numeric.}
}
\value{
The \code{variable} argument also allows using \code{variable = 1} for categorical variable and \code{variable = 2} for continuous variable.
}
\description{
Constructor of object of leem class
}
\examples{
# Example 1
library(leem)
x <- rbinom(36, 10, 0.6)
new_leem(x, variable = 1)

# Example 2 (Pipe operator)
rnorm(36, 100, 4) |> new_leem(variable = 2)

}
