% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.R
\name{PF}
\alias{PF}
\title{Probability function}
\usage{
PF(
  x,
  dist = "binomial",
  rounding = 5,
  porcentage = FALSE,
  gui = "plot",
  main = NULL,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{represents the assumed value of the random variable X. The \code{x} argument is a value of support of X. See Details.}

\item{dist}{distribution to use. The default is \code{'binomial'}. Options: \code{'binomial'} and \code{'poisson'}. The other discrete distributions are being implemented.}

\item{rounding}{numerical; it represents the number of decimals for calculating the probability.}

\item{porcentage}{logical; if \code{FALSE} (default), the result in decimal. Otherwise, probability is given in percentage.}

\item{gui}{default is \code{'plot'}; it graphically displays the result of the probability. Others options are: \code{'none'}, \code{'rstudio'} or \code{'tcltk'}.}

\item{plot}{logical; if \code{TRUE} (default), the result will be shown along with its graphical representation; otherwise, only the result will be shown.}

\item{...}{additional arguments according to the chosen distribution.}
}
\value{
\code{PF} returns the probability and its graphical representation. The result can be given as a percentage or not.
}
\description{
\code{PF} This function computes the value of the probability function for the discrete random variables
}
\details{
The standard distribution to be computed will be the Binomial distribution (\code{dist='binomial'}). Once you forget the function arguments, don't worry, they will be requested via the console. This means if the user does not know the parameters of the distribution, the \code{pf()} function takes care of asking the user.
\itemize{
\item If \code{dist='binomial'}: \eqn{X \sim Bin(n, p)}
\itemize{
\item additional arguments: \code{size} \eqn{(n)} and \code{prob} \eqn{(p)};
\item Suport of X: \eqn{S_X=\{0,1,..., n\}}; this represents the values for the \code{x} argument;
}
}
}
\examples{
# Loading package
library(leem)
# Example 1 - Binomial distribution
\dontrun{
PF(2, dist = "binomial", size = 10, prob = 0.8)
}
}
