% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{polyfreq}
\alias{polyfreq}
\title{Frequency polygon Graph}
\usage{
polyfreq(x, ...)
}
\arguments{
\item{x}{R object (list) of class leem. Use \code{new_leem()} function.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function that plots the frequency polygon curve.
}
\examples{
# Example 1
library(leem)
rnorm(36, 100, 50) |> new_leem(variable = "continuous") |> tabfreq() |> polyfreq()
}
