% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propofnormal.R
\name{propofnormal}
\alias{propofnormal}
\title{Properties of the normal distribution}
\usage{
propofnormal(col = "lightblue2", type = 1)
}
\arguments{
\item{col}{color type.}

\item{type}{numerical. Type of properties. Options: \code{1}, \code{2}, \code{3}, \code{4} and \code{5}. Default \code{code = 1}.}
}
\description{
Graphically it is possible to observe some properties of the normal distribution
}
