% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_prob.R
\name{Regions_of_probability}
\alias{Regions_of_probability}
\alias{\%>x>\%}
\alias{\%>X>\%}
\alias{\%<x<\%}
\alias{\%<X<\%}
\alias{\%>=x>=\%}
\alias{\%>=X>=\%}
\alias{\%<=x<=\%}
\alias{\%<=X<=\%}
\alias{\%>=x>\%}
\alias{\%>=X>\%}
\alias{\%>x>=\%}
\alias{\%>X>=\%}
\alias{\%<=x<\%}
\alias{\%<=X<\%}
\alias{\%<x<=\%}
\alias{\%<X<=\%}
\title{Regions of probability}
\usage{
a \%>x>\% b

a \%>X>\% b

a \%<x<\% b

a \%<X<\% b

a \%>=x>=\% b

a \%>=X>=\% b

a \%<=x<=\% b

a \%<=X<=\% b

a \%>=x>\% b

a \%>=X>\% b

a \%>x>=\% b

a \%>X>=\% b

a \%<=x<\% b

a \%<=X<\% b

a \%<x<=\% b

a \%<X<=\% b
}
\arguments{
\item{a}{scalar. when referring to a discrete random variable, use the syntax "\code{L}" after the number.}

\item{b}{scalar. when referring to a discrete random variable, use the syntax "\code{L}" after the number.}
}
\value{
A vector of lenght 2.
}
\description{
These binary operators return a vector of
length 2, describing the desired probability
region.
}
\examples{
# Example 1 - Discrete
2L \%>x>\% 5L
2L \%>X>\% 5L
2L \%<X<\% 5L
2L \%<x<\% 5L
2L \%>=X>=\% 5L
2L \%>=x>=\% 5L
2L \%<=X<=\% 5L
2L \%<=x<=\% 5L
2L \%>=X>\% 5L
2L \%>=x>\% 5L
2L \%>x>=\% 5L
2L \%>X>=\% 5L
2L \%<=X<\% 5L
2L \%<=x<\% 5L
2L \%<X<=\% 5L
2L \%<x<=\% 5L

# Example 2 - Continuous
2 \%>x>\% 5
2 \%>X>\% 5
2 \%<X<\% 5
2 \%<x<\% 5
2 \%>=X>=\% 5
2 \%>=x>=\% 5
2 \%<=X<=\% 5
2 \%<=x<=\% 5
2 \%>=X>\% 5
2 \%>=x>\% 5
2 \%>x>=\% 5
2 \%>X>=\% 5
2 \%<=X<\% 5
2 \%<=x<\% 5
2 \%<X<=\% 5
2 \%<x<=\% 5

}
