% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showboxplot.R
\name{showboxplot}
\alias{showboxplot}
\title{Undertanding a box plot}
\usage{
showboxplot(horizontal = FALSE, col = rgb(0, 175, 239, maxColorValue = 255))
}
\arguments{
\item{horizontal}{Logical argument indicating if the boxplots should be horizontal; default \code{FALSE} means vertical boxes.}

\item{col}{Character vector. Default \code{col = rgb(0, 175, 239, maxColorValue = 255)}.}
}
\description{
Detailing of a box plot, showing the main information contained in this type of graph
}
\examples{
library(leem)
# Example 1
showboxplot()

}
