% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showci.R
\name{showci}
\alias{showci}
\title{Understanding the Confiance Indice}
\usage{
showci(dist = "normal", ci = "two.sided", main = NULL)
}
\arguments{
\item{dist}{Parameter to indicate the distribution of the graphic, fixed for now.}

\item{ci}{Parameter to indicate the region of the confiance indice.}

\item{main}{Parameter to indicate the title of the graphic.}
}
\description{
Detailing the confiance indice plot, showing the main information contained in this type of graph.
}
\examples{
library(leem)
# Example 1
showci()

}
