% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showkur.R
\name{showkur}
\alias{showkur}
\title{Plot of interpretation about Kurtosis}
\usage{
showkur()
}
\value{
\code{showkur} returns a plot with the kurtosis characteristics.
}
\description{
\code{showkur} Interpretation of kutosis
}
\examples{
# Loading package
library(leem)
\dontrun{
showkur()
}

}
