% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showpar.R
\name{showpar}
\alias{showpar}
\title{Interpretation of location and scale parameters}
\usage{
showpar(gui = "rstudio")
}
\arguments{
\item{gui}{character argument. The options are: \code{"rstudio"} (default) and \code{"tcltk"}.}
}
\value{
\code{showpar} returns an interactive plot.
}
\description{
\code{showpar} Function that exemplifies the interpretation of location and scale parameters
}
\details{
The result of the \code{showpar()} call will interactively present a plot of the normal distribution showing the behavior of the location and scale parameters via RStudio. For \code{showpar(gui = "tcltk")} the result will be displayed in a tcltk interface.
}
\examples{
# Loading package
library(leem)
\dontrun{
showpar()
}

}
