% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showsf.R
\name{showsf}
\alias{showsf}
\title{Survival Function Properties}
\usage{
showsf(variable = "discrete", prop = NULL)
}
\arguments{
\item{variable}{Variabe type. Defaults \code{variable = "discrete"}. The options are: \code{discrete} or "\code{1}" and \code{continuous} or "\code{2}".}

\item{prop}{Properties for survival function. See Details.}
}
\description{
Graphic presentation of properties for survival function
}
\details{
\itemize{
\item \code{prop = 1}: \eqn{\lim_{x\rightarrow\infty}S_X(x)=0} and \eqn{\lim_{x\rightarrow -\infty}S_X(x)=1};
\item \code{prop = 2}: \eqn{S_X(x)\geq S_X(y), ~ x\leq y~\forall x,y \in \mathbb{R}};
\item \code{prop = 3}: \eqn{\lim_{x_n\downarrow x}S_X(x_n)\downarrow S_X(x)}.
}
}
\examples{
library(leem)
# Example 1: Plot of survival function
showsf()
}
