% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_of_hypothesis.R
\name{th}
\alias{th}
\title{Test of hypothesis}
\usage{
th(
  x,
  y = NULL,
  test = "ztest",
  h0,
  prop = FALSE,
  p,
  pa,
  alternative = c("two.sided", "L", "less", "greater", "G"),
  alpha = 0.05,
  exact = TRUE,
  correct = FALSE,
  paired = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{x}{R object. See in details.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{test}{character value. The options are: \code{"ttest"}, \code{"ztest"}, \code{"ptest"}, \code{"chitest"}, \code{"ftest"}, \code{"anova"}, \code{"friedman"}, \code{"kruskal"}, \code{"mann whitney"}.}

\item{h0}{numeric value. The hypothesized parameter.}

\item{prop}{a logical indicating whether you want to use the proportion test of not. Default is \code{prop=FALSE}.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{alpha}{significance level of the test}

\item{exact}{a logical indicating whether you want to use the exact test or not. Default is \code{exact=TRUE}.}

\item{correct}{a logical indicating whether Yates' continuity correction should be applied where possible. This argument must be used when \code{exact = FALSE}.}

\item{paired}{a logical indicating whether you want a paired t-test. Valid only for \code{test="ttest"}.}

\item{plot}{a logical indicating whether you want a graph indicating the regions of rejection or not of the null hypothesis, as well as the test decision.}
}
\description{
Performs hypothesis testing for various parameters of one or more populations
}
