% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Insert measures of position in plot}
\usage{
insert(dados, ...)

## Leem S3 method:
insert(x, type = "black", lcol, tcol = lcol, acol = lcol, parrow = 0.5,
       larrow = 0.2, ptext = 0.6, side = "right", lwd = 2, lwdarrow = lwd)

## Default S3 method:
insert(x)
}
\arguments{
\item{x}{R object (list) of class leem. Use \code{new_leem()} function.}

\item{type}{Type of measure of position. The default is \code{type = "mean"}. Other options: \code{"median"}, \code{"mode"} or \code{"all"}.}

\item{lcol}{Vertical line color type. The default is \code{lpcol = "black"}. This argument must be the same length as the \code{type} argument.}

\item{tcol}{Text color type. The default is \code{tcol = lcol}.}

\item{acol}{Arrow color type. The default is \code{acol = lcol}.}

\item{parrow}{Text and arrow height. The default is \code{parrow = 0.5}. This argument must be the same length as the \code{type} argument.}

\item{larrow}{Text and arrow length. The default is \code{larrow = 0.3}.}

\item{ptext}{Distance between lines of text. The default is \code{ptext = 0}.}

\item{side}{Side to insert the text. The default is \code{side = "right"}. This argument must be the same length as the \code{type} argument.}

\item{lwd}{numeric argument. The vertical line width. The default is  \code{lwd = 2}.}

\item{lwdarrow}{numeric argument. The arrow width. The default is  \code{lwdarrow = lwd}.}
}
\value{
The result of \code{tabfreq()} is a list. This list has two elements: \code{table} and \code{statistics}. The first is the data frequency table, and the second represents some useful statistics for methods of leem class.
}
\description{
Generic function that allows inserting measures of position in plots
}
\examples{
# Example 1
library(leem)
set.seed(10)
rnorm(36, 100, 50) |>
 new_leem(variable = "continuous") |>
 tabfreq() |>
 hist() |>
 insert(
  lcol = "black",
  tcol = "purple",
  acol = "brown",
  parrow = 0.6,
  larrow = 0.6,
  ptext = 0.4,
  side = "left",
  lwd = 2,
  lwdarrow = 4
 )
}
