% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{P}
\alias{P}
\title{Cumulative distribution function}
\usage{
P(
  q,
  dist = "normal",
  lower.tail = TRUE,
  rounding = 5,
  porcentage = FALSE,
  gui = "plot",
  main = NULL,
  ...
)
}
\arguments{
\item{q}{numerical. The \code{q} argument can have length 1 or 2. See Details.}

\item{dist}{distribution to use. The default is \code{'normal'}. Options: \code{'normal'} (Normal distribution), \code{'t-student'} (Student's distribution), \code{chisq} (Chi-squared distribution), \code{'f'} (F distribution), \code{'gumbel'} (Gumbel distribution), \code{'beta'} (Beta distribution), \code{'exp'} (Exponential distribution), \code{'gamma'} (Gamma distribution), \code{'cauchy'} (Cauchy distribution), \code{'logis'} (Logistic distribution), \code{'lnormal'} (Log-normal distriburion), \code{'tukey'} (Tukey distribution), \code{'weibull'} (Weibull distribution), \code{'poisson'} (Poisson distribution), \code{'binomial'} (Binomial distribution), \code{'nbinom'} (Negative binomial distribution), \code{'hyper'} (Hypergeometric distribution), \code{'geom'} (Geometric distribution), \code{'unif'} (Uniform distribution), \code{'wilcox'} (Wilcoxon distribution), and \code{'signrank'} (Sign Rank distribution).}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}. This argument is valid only if \code{q} has length 1.}

\item{rounding}{numerical; it represents the number of decimals for calculating the probability.}

\item{porcentage}{logical; if \code{FALSE} (default), the result in decimal. Otherwise, probability is given in percentage.}

\item{gui}{default is \code{'plot'}; it graphically displays the result of the probability. Others options are: \code{'none'}, \code{'rstudio'} or \code{'tcltk'}.}

\item{...}{additional arguments according to the chosen distribution.}
}
\value{
\code{P} returns the probability and its graphical representation. The result can be given as a percentage or not.
}
\description{
\code{P} Compute the cumulative distribution function for multiple distributions
}
\details{
The argument that can have length 2, when we use the functions that give us the probability regions, given by: \code{\%<X<\%} (\eqn{(P[a < X < b])}), \code{\%<=X<\%} \eqn{(P[a \leq X < b])}, \code{\%<X<=\%} \eqn{(P[a < X \leq b])}, \code{\%<=X<=\%} \eqn{(P[a \leq X \leq b])}, \code{\%>X>\%} \eqn{(P[X < a] + P[X > b])}, \code{\%>X=>\%} \eqn{(P[X < a] + P[X \geq b])}, \code{\%>X=>\%} \eqn{(P[X < a] + P[X \geq b])} and \code{\%>=X=>\%} \eqn{(P[X < a] + P[X \geq b])}, considering \eqn{a,~b \in \mathbb{R}} and \eqn{b > a}.
The additional arguments represent the parameters of the distributions, that is:
\itemize{
\item \code{dist = "normal"}: \code{mu} argument (\eqn{\mu}) represents the averages and \code{sd} argument (\eqn{\sigma}) represents the standard deviation parameters. The PDF is
}

\deqn{\displaystyle f(x)={\frac {1}{\sigma {\sqrt {2\pi }}}}e^{-{\frac {1}{2}}\left({\frac {x-\mu }{\sigma }}\right)^{2}}, \quad -\infty \geq x \geq \infty, \quad -\infty \geq \mu \geq \infty, \quad \sigma > 0;}
\itemize{
\item \code{dist = "t-student"}: \code{nu} argument (\eqn{\nu}) represents the degrees of freedom parameter. The PDF is
}

\deqn{\displaystyle f(t)={\frac {\ \Gamma \left({\frac {\ \nu +1\ }{2}}\right)\ }{\ {\sqrt {\pi \ \nu \ }}\ \Gamma \left({\frac {\nu }{2}}\right)}}\left(\ 1+{\frac {~t^{2}\ }{\nu }}\ \right)^{-(\nu +1)/2}
\quad -\infty < t < \infty,~ \nu > 0,}
where \eqn{\G0amma()} is gamma function;
\itemize{
\item \code{dist = "chisq"}: \code{nu} argument (\eqn{\nu}) represents the degrees of freedom parameter. The PDF is
}

\deqn{\displaystyle f(x)=\dfrac {x^{\nu/2-1}e^{-x/2}}{2^{\nu/2}\Gamma \left({\frac {\nu}{2}}\right)}, \quad x > 0, \quad \nu > 0,}
where \eqn{\Gamma()} is gamma function;
}
\examples{
# Loading package
library(leem)
# Example 1 - t-Student distribution
\dontrun{
P(q = 2, dist = "t-student", df = 10)
P(q = 2, dist = "t-student", df = 10, gui = 'rstudio')
P(q = 2, dist = "t-student", df = 10, gui = 'tcltk')
P(-1 \%<X<\% 1, dist = "t-student", df = 10)
}
# Example 2 - Normal distribution
P(-2,  dist = "normal", mean = 3, sd = 2, main = expression(f(x) == (1 / sqrt(n * sigma^2)) * exp(-1/2 * (x - mu)^2/sigma^2)))
}
