% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{P}
\alias{P}
\title{Cumulative distribution function}
\usage{
P(
  q,
  dist = "normal",
  lower.tail = TRUE,
  rounding = 5,
  porcentage = FALSE,
  gui = "plot",
  main = NULL,
  ...
)
}
\arguments{
\item{q}{numerical. The \code{q} argument can have length 1 or 2. See Details.}

\item{dist}{distribution to use. The default is \code{'normal'}. Options: \code{'normal'} (Normal distribution), \code{'t-student'} (Student's distribution),
\code{'chisq'} (Chi-squared distribution), \code{'f'} (F distribution), \code{'gumbel'} (Gumbel distribution), \code{'beta'} (Beta distribution),
\code{'exp'} (Exponential distribution), \code{'gamma'} (Gamma distribution), \code{'cauchy'} (Cauchy distribution), \code{'logis'}
(Logistic distribution), \code{'lnormal'} (Log-normal distriburion), \code{'tukey'} (Tukey distribution), \code{'weibull'} (Weibull distribution),
\code{'poisson'} (Poisson distribution), \code{'binomial'} (Binomial distribution), \code{'nbinom'} (Negative binomial distribution), \code{'hyper'}
(Hypergeometric distribution), \code{'geom'} (Geometric distribution), \code{'unif'} (Uniform distribution), \code{'wilcox'} (Wilcoxon distribution),
and \code{'signrank'} (Sign Rank distribution).}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}. This argument is valid only if \code{q} has length 1.}

\item{rounding}{numerical; it represents the number of decimals for calculating the probability.}

\item{porcentage}{logical; if \code{FALSE} (default), the result in decimal. Otherwise, probability is given in percentage.}

\item{gui}{default is \code{'plot'}; it graphically displays the result of the probability. Others options are: \code{'none'}, \code{'rstudio'} or \code{'tcltk'}.}

\item{...}{additional arguments according to the chosen distribution.}
}
\value{
\code{P} returns the probability and its graphical representation. The result can be given as a percentage or not.
}
\description{
\code{P} Compute the cumulative distribution function for multiple distributions
}
\details{
The argument that can have length 2, when we use the functions that give us
the probability regions, given by: \code{\%<X<\%} (\eqn{(P[a < X < b])}), \code{\%<=X<\%}
\eqn{(P[a \leq X < b])}, \code{\%<X<=\%} \eqn{(P[a < X \leq b])}, \code{\%<=X<=\%}
\eqn{(P[a \leq X \leq b])}, \code{\%>X>\%} \eqn{(P[X < a] + P[X > b])}, \code{\%>X=>\%}
\eqn{(P[X < a] + P[X \geq b])}, \code{\%>X=>\%} \eqn{(P[X < a] + P[X \geq b])} and
\code{\%>=X=>\%} \eqn{(P[X < a] + P[X \geq b])}, considering \eqn{a,~b \in \mathbb{R}} and \eqn{b > a}.
The additional arguments represent the parameters of the distributions, that is:
\itemize{
\item \code{dist = "normal"}: \code{mu} argument (\eqn{\mu}) represents the averages and \code{sd} argument (\eqn{\sigma}) represents the standard deviation parameters. The PDF is
}

\deqn{\displaystyle f(x)={\frac {1}{\sigma {\sqrt {2\pi }}}}e^{-{\frac {1}{2}}\left({\frac {x-\mu }{\sigma }}\right)^{2}}, \quad -\infty \geq x \geq \infty, \quad -\infty \geq \mu \geq \infty, \quad \sigma > 0;}
\itemize{
\item \code{dist = "t-student"}: \code{df} argument (\eqn{\nu}) represents the degrees of freedom parameter. The PDF is
}

\deqn{\displaystyle f(t)={\frac {\ \Gamma \left({\frac {\ \nu +1\ }{2}}\right)\ }{\ {\sqrt {\pi \ \nu \ }}\ \Gamma \left({\frac {\nu }{2}}\right)}}\left(\ 1+{\frac {~t^{2}\ }{\nu }}\ \right)^{-(\nu +1)/2}
\quad -\infty < t < \infty,~ \nu > 0,}
where \eqn{\Gamma} is the gamma function;
\itemize{
\item \code{dist = "chisq"}: \code{df} argument (\eqn{\nu}) represents the degrees of freedom parameter. The PDF is
}

\deqn{\displaystyle f(x)=\dfrac {x^{\nu/2-1}e^{-x/2}}{2^{\nu/2}\Gamma \left({\frac {\nu}{2}}\right)}, \quad x > 0, \quad \nu > 0,}
where \eqn{\Gamma} is the gamma function;
\itemize{
\item \code{dist = "f"}: \code{df1} and \code{df2} arguments (\eqn{\nu_1} and \eqn{\nu_2}) represents the degrees of freedom parameters. The PDF is
}

\deqn{f(x) = \frac{\Gamma((\nu_1 + \nu_2) / 2)}{\Gamma(\nu_1 / 2)\Gamma(\nu_2 / 2)}\left(\frac{\nu_1}{\nu_2}\right)^{\nu_1 / 2} \times \frac{x^{(\nu_1 - 2) / 2}}{[1 + (\nu_1 / \nu_2)x]^{(\nu_1 + \nu_2)/2}}, \quad x > 0,\quad \nu_1,~\nu_2 > 0,}
where \eqn{\Gamma} is the gamma function;
\itemize{
\item \code{dist = "gumbel"}: \code{scale} and \code{location} arguments (\eqn{\beta} and \eqn{\mu}) represents the scale and location parameters. The PDF is
}

\deqn{f(x) = \frac{1}{\beta}e^{-(z + e^{-z}) / \beta}, \quad z = \frac{x - \mu}{\beta}, \quad x \in \mathbb{R}, \quad \mu \in \mathbb{R}, \quad \beta > 0;}
\itemize{
\item \code{dist = "beta"}: \code{alpha} and \code{beta} arguments represents the \eqn{alpha} and \eqn{\beta} parameters. The PDF is
}

\deqn{f(x) = \frac{1}{B(\alpha,\beta)}x^{\alpha - 1}(1 - x)^{\beta - 1}, \quad x \in [0, 1], \quad \alpha, \beta > 0,}
where \eqn{\displaystyle \mathrm {B} (\alpha ,\beta )={\frac {\Gamma (\alpha )\Gamma (\beta )}{\Gamma (\alpha +\beta )}}} and \eqn{\Gamma} is the gamma function;
\itemize{
\item \code{dist = "exp"}: \code{rate} argument represents the \eqn{\lambda} parameter. The PDF is
}

\deqn{f(x) = \lambda e^{\lambda x}, \quad x > 0, \quad \lambda > 0;}
\itemize{
\item code{dist = "gamma"}: \code{shape} and \code{rate} arguments represents the \eqn{\alpha} and \eqn{\beta} parameters. The PDF is
}

\deqn{f(x) = \frac{\beta^{\alpha}}{\Gamma(\alpha)}x^{\alpha - 1}e^{-\beta x},\quad x > 0, \quad \alpha, \beta > 0,}
where \eqn{\Gamma} is the gamma function;
\itemize{
\item \code{dist = "cauchy"}: \code{location} and \code{scale} arguments represents \eqn{\alpha} and \eqn{\beta} parameters. The PDF is
}

\deqn{f(x) = \displaystyle \frac{1}{\pi \beta \left[ 1 + \left( \frac{x - \alpha}{\beta}\right)^2\right]},\quad x \in \mathbb{R}, \quad \alpha \in \mathbb{R}, \quad \beta > 0;}
\itemize{
\item \code{dist = "logis"}: \code{location} and \code{scale} arguments represents \eqn{\mu} and \eqn{\sigma} parameters. The PDF is
}

\deqn{f(x)= \displaystyle \frac {e^{-(x-\mu )/\sigma}}{\sigma\left(1+e^{-(x-\mu )/\sigma}\right)^{2}}, \quad x \in \mathbb{R}, \quad \mu \in \mathbb{R}, \quad \sigma > 0;}
\itemize{
\item \code{dist = "lnormal"}: \code{mean} and \code{sd} arguments represents \eqn{\mu} and \eqn{\sigma} parameters. The PDF is
}

\deqn{f(x)= \displaystyle {\frac {1}{\ x\ \sigma {\sqrt {2\ \pi \ }}\ }}\exp \left(-{\frac {\ (\ln x-\mu )^{2}\ }{2\ \sigma ^{2}}}\right), \quad x \in \mathbb{R}, \quad \mu \in \mathbb{R}, \quad \sigma > 0;}
\itemize{
\item \code{dist = "tukey"}: \code{mean} and \code{sd} arguments represents \eqn{\mu} and \eqn{\sigma} parameters. The PDF is
}
}
\examples{
# Loading package
library(leem)
# Example 1 - t-Student distribution
\dontrun{
P(q = 2, dist = "t-student", df = 10)
P(q = 2, dist = "t-student", df = 10, gui = 'rstudio')
P(q = 2, dist = "t-student", df = 10, gui = 'tcltk')
P(-1 \%<X<\% 1, dist = "t-student", df = 10)
}
# Example 2 - Normal distribution
P(-2,  dist = "normal", mean = 3, sd = 2, main = expression(f(x) == (1 / sqrt(n * sigma^2)) * exp(-1/2 * (x - mu)^2/sigma^2)))
}
