% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{p}
\alias{p}
\title{Cumulative distribution function}
\usage{
p(
  q,
  dist = "t-student",
  lower.tail = TRUE,
  rounding = 4,
  porcentage = FALSE,
  gui = "plot",
  ...
)
}
\arguments{
\item{q}{quantile. The \code{q} argument can have length 1 or 2. See Details.}

\item{dist}{distribution to use. The default is 't-student'. Options: \code{'normal'}, \code{'t-student'}, \code{'binomial'}, \code{'poisson'}, and ....}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}. This argument is valid only if \code{q} has length 1.}

\item{rounding}{numerical; it represents the number of decimals for calculating the probability.}

\item{porcentage}{logical; if \code{FALSE} (default), the result in decimal. Otherwise, probability is given in percentage.}

\item{gui}{default is \code{'plot'}; it graphically displays the result of the probability. Others options are: \code{'rstudio'} or \code{'tcltk'}.}

\item{...}{additional arguments according to the chosen distribution.}
}
\value{
\code{p} returns the probability and its graphical representation. The result can be given as a percentage or not.
}
\description{
\code{p} Cumulative distribution function for multiple distributions
}
\details{
The argument that can have length 2, when we use the functions that give us the probability regions, given by: \code{\%<X<\%}, \code{\%<=X<\%}, \code{\%<X<=\%}, \code{\%<=X<=\%}, \code{\%>X>\%}, \code{\%>X=>\%}, \code{\%>X=>\%} and \code{\%>=X=>\%}.
The additional arguments represent the parameters of the distributions, that is:

- \code{dist = "t-student"}: \code{nu}
}
\examples{
# Loading package
library(leem)
\dontrun{
p(q = 2, dist = "t-student", df = 10)
p(q = 2, dist = "t-student", df = 10, gui = 'rstudio')
p(q = 2, dist = "t-student", df = 10, gui = 'tcltk')
p(-1 \%<X<\% 1, dist = "t-student", df = 10)
}
}
