% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{Q}
\alias{Q}
\title{Quantile distribution function.}
\usage{
Q(
  p,
  dist = "t-student",
  lower.tail = TRUE,
  rounding = 2,
  gui = "plot",
  mfrow = c(1, 2),
  type = "both",
  ...
)
}
\arguments{
\item{p}{probability. The \code{p} argument need have length 1 and value lower then 1.}

\item{dist}{distribution to use. The default is \code{'t-student'}. Options: \code{'normal'}, \code{'t-student'}, \code{'gumbel'}, \code{'binomial'}, \code{'poisson'}, and ....}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}. This argument is valid only if \code{q} has length 1.}

\item{rounding}{numerical; it represents the number of decimals for calculating the probability.}

\item{gui}{default is \code{'plot'}; it graphically displays the result of the probability. Others options are: \code{'plot'} and....}

\item{...}{additional arguments according to the chosen distribution.}
}
\value{
\code{Q} returns the point and its graphical representation. The result is give by a integer number.
}
\description{
\code{Q} Quantile function for multiple distributions.
}
\details{
The expression of quantile function is given by:
\deqn{
Q(p)=\inf {x\in \mathbb{R}: p \leq F(x)},
}
where \code{p} is the first argument of \code{Q()} and \code{x} its return value;
}
\examples{
# Loading package
library(leem)
\dontrun{
Q(p = 0.8, dist = "normal", mean = 200, sd=30)
}
}
