% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showskew.R
\name{showskew}
\alias{showskew}
\title{Plot of interpretation about skewsness}
\usage{
showskew(mpos = FALSE)
}
\arguments{
\item{measures}{shows the measures of position or not (default \code{FALSE}).}
}
\value{
\code{showskew} returns a plot with the skewsness characteristics.
}
\description{
\code{showskew} Interpretation of asymmetry based on frequency distributions
}
\examples{
# Loading package
library(leem)
\dontrun{
showskew()
}

}
