% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showtabnormal.R
\name{showtabnormal}
\alias{showtabnormal}
\title{Understanding the Ztable.}
\usage{
showtabnormal(z)
}
\arguments{
\item{z}{Parameter for lacate the z value on the table; default \code{NULL} means error if not insert a number.}
}
\description{
Detailing of the Ztable, showing the main information contained in this type of table.
}
\examples{
library(leem)
# Example 1
showtabnormal(zvalue)
}
